Add lint fixes to test package
