+++ test/e2e/capx_quick_start_test.go
@@ -1,5 +1,4 @@
 //go:build e2e
-// +build e2e
 
 /*
 Copyright 2020 The Kubernetes Authors.
