+++ test/e2e/clusterctl_upgrade_test.go
@@ -1,8 +1,7 @@
 //go:build e2e
-// +build e2e
 
 /*
-Copyright 2020 The Kubernetes Authors.
+Copyright 2023 Nutanix.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
